Move grpc examples to app-examples module
