+++ app-examples/src/main/java/net/spals/appbuilder/app/examples/grpc/minimal/MinimalGrpcWebApp.java
@@ -1,4 +1,4 @@
-package net.spals.appbuilder.app.grpc.minimal;
+package net.spals.appbuilder.app.examples.grpc.minimal;
 
 import net.spals.appbuilder.app.grpc.GrpcWebApp;
