+++ app-examples/src/main/java/net/spals/appbuilder/app/examples/grpc/rest/web/UserGrpcServiceV2.java
@@ -1,46 +1,44 @@
-package net.spals.appbuilder.app.grpc.rest.web;
+package net.spals.appbuilder.app.examples.grpc.rest.web;
 
 import com.google.inject.Inject;
-import com.google.protobuf.Descriptors;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import io.grpc.stub.StreamObserver;
 import net.spals.appbuilder.annotations.service.AutoBindSingleton;
-import net.spals.appbuilder.app.grpc.rest.*;
+import net.spals.appbuilder.app.examples.grpc.rest.*;
 import net.spals.appbuilder.mapstore.core.MapStore;
 import net.spals.appbuilder.mapstore.core.model.MapStoreKey;
 
-import java.util.AbstractMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 /**
- * A sample gRPC service using proto3
+ * A sample gRPC service using proto2
  * which has a complementary RESTful interface.
  *
  * @author tkral
  */
 @AutoBindSingleton
-public class UserGrpcServiceV3 extends UserServiceV3Grpc.UserServiceV3ImplBase {
+public class UserGrpcServiceV2 extends UserServiceV2Grpc.UserServiceV2ImplBase {
 
     private final MapStore mapStore;
 
     @Inject
-    UserGrpcServiceV3(final MapStore mapStore) {
+    UserGrpcServiceV2(final MapStore mapStore) {
         this.mapStore = mapStore;
     }
 
     @Override
-    public void deleteUserV3(
-        final DeleteUserRequestV3 request,
-        final StreamObserver<DeleteUserResponseV3> responseObserver
+    public void deleteUserV2(
+        final DeleteUserRequestV2 request,
+        final StreamObserver<DeleteUserResponseV2> responseObserver
     ) {
         final MapStoreKey userRecordKey = createUserRecordKey(request.getId());
         mapStore.deleteItem("users", userRecordKey);
 
         responseObserver.onNext(
-            DeleteUserResponseV3.newBuilder()
+            DeleteUserResponseV2.newBuilder()
                 .setId(request.getId())
                 .build()
         );
@@ -48,19 +46,19 @@ public class UserGrpcServiceV3 extends UserServiceV3Grpc.UserServiceV3ImplBase {
     }
 
     @Override
-    public void getUserV3(
-        final GetUserRequestV3 request,
-        final StreamObserver<GetUserResponseV3> responseObserver
+    public void getUserV2(
+        final GetUserRequestV2 request,
+        final StreamObserver<GetUserResponseV2> responseObserver
     ) {
         final MapStoreKey userRecordKey = createUserRecordKey(request.getId());
         final Optional<Map<String, Object>> userRecordMap = mapStore.getItem("users", userRecordKey);
 
-        final GetUserResponseV3.Builder responseBuilder = GetUserResponseV3.newBuilder();
+        final GetUserResponseV2.Builder responseBuilder = GetUserResponseV2.newBuilder();
         userRecordMap.ifPresent(map -> {
             map.entrySet().stream()
-                .filter(entry -> GetUserResponseV3.getDescriptor().findFieldByName(entry.getKey()) != null)
+                .filter(entry -> GetUserResponseV2.getDescriptor().findFieldByName(entry.getKey()) != null)
                 .forEach(entry -> {
-                    final FieldDescriptor fieldDescriptor = GetUserResponseV3.getDescriptor().findFieldByName(entry.getKey());
+                    final FieldDescriptor fieldDescriptor = GetUserResponseV2.getDescriptor().findFieldByName(entry.getKey());
                     responseBuilder.setField(fieldDescriptor, entry.getValue());
                 });
         });
@@ -70,14 +68,14 @@ public class UserGrpcServiceV3 extends UserServiceV3Grpc.UserServiceV3ImplBase {
     }
 
     @Override
-    public void postUserV3(
-        final PostUserRequestV3 request,
-        final StreamObserver<PostUserResponseV3> responseObserver
+    public void postUserV2(
+        final PostUserRequestV2 request,
+        final StreamObserver<PostUserResponseV2> responseObserver
     ) {
         final String userId = UUID.randomUUID().toString();
 
         final MapStoreKey userRecordKey = createUserRecordKey(userId);
-        final UserRecordV3 userRecord = UserRecordV3.newBuilder()
+        final UserRecordV2 userRecord = UserRecordV2.newBuilder()
             .setId(userId)
             .setName(request.getName())
             .build();
@@ -87,7 +85,7 @@ public class UserGrpcServiceV3 extends UserServiceV3Grpc.UserServiceV3ImplBase {
         mapStore.putItem("users", userRecordKey, userRecordMap);
 
         responseObserver.onNext(
-            PostUserResponseV3.newBuilder()
+            PostUserResponseV2.newBuilder()
                 .setId(userId)
                 .setName(request.getName())
                 .build()
