+++ app-examples/src/main/java/net/spals/appbuilder/app/examples/grpc/sample/SampleGrpcCustomSet.java
@@ -1,4 +1,4 @@
-package net.spals.appbuilder.app.grpc.sample;
+package net.spals.appbuilder.app.examples.grpc.sample;
 
 /**
  * A multi service definition for {@link SampleGrpcWebApp}
