+++ app-examples/src/main/java/net/spals/appbuilder/app/examples/grpc/sample/SampleGrpcGuiceModule.java
@@ -1,4 +1,4 @@
-package net.spals.appbuilder.app.grpc.sample;
+package net.spals.appbuilder.app.examples.grpc.sample;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Module;
