+++ app-examples/src/main/java/net/spals/appbuilder/app/examples/grpc/sample/web/SampleGrpcServiceV3.java
@@ -1,11 +1,11 @@
-package net.spals.appbuilder.app.grpc.sample.web;
+package net.spals.appbuilder.app.examples.grpc.sample.web;
 
 import com.google.inject.Inject;
 import io.grpc.stub.StreamObserver;
 import net.spals.appbuilder.annotations.service.AutoBindSingleton;
-import net.spals.appbuilder.app.grpc.sample.SampleRequestV3;
-import net.spals.appbuilder.app.grpc.sample.SampleResponseV3;
-import net.spals.appbuilder.app.grpc.sample.SampleServiceV3Grpc;
+import net.spals.appbuilder.app.examples.grpc.sample.SampleRequestV3;
+import net.spals.appbuilder.app.examples.grpc.sample.SampleResponseV3;
+import net.spals.appbuilder.app.examples.grpc.sample.SampleServiceV3Grpc;
 
 /**
  * A sample gRPC service using proto3.
