+++ app-examples/src/main/protobuf/rest_routes_v3.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 option java_multiple_files = true;
 // Should be included in the service packages scanned by ServiceScan
-option java_package = "net.spals.appbuilder.app.grpc.rest";
+option java_package = "net.spals.appbuilder.app.examples.grpc.rest";
 option java_outer_classname = "RestServerV3Proto";
 
 import "google/api/annotations.proto";
