+++ app-examples/src/main/protobuf/sample_routes_v2.proto
@@ -1,7 +1,7 @@
 syntax = "proto2";
 
 option java_multiple_files = true;
-option java_package = "net.spals.appbuilder.app.grpc.sample";
+option java_package = "net.spals.appbuilder.app.examples.grpc.sample";
 option java_outer_classname = "SampleServerV2Proto";
 
 service SampleServiceV2 {
