+++ app-examples/src/main/resources/config/rest-grpc-service.conf
