+++ app-grpc-test/pom.xml
@@ -19,14 +19,6 @@
             <artifactId>jackson-jaxrs-json-provider</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.google.protobuf</groupId>
-            <artifactId>protobuf-java</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.xorlev.grpc-jersey</groupId>
-            <artifactId>jersey-rpc-support</artifactId>
-        </dependency>
-        <dependency>
             <groupId>eu.codearte.catch-exception</groupId>
             <artifactId>catch-exception</artifactId>
         </dependency>
@@ -35,44 +27,8 @@
             <artifactId>catch-throwable</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.grpc</groupId>
-            <artifactId>grpc-protobuf</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.grpc</groupId>
-            <artifactId>grpc-stub</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.opentracing</groupId>
-            <artifactId>opentracing-mock</artifactId>
-        </dependency>
-        <dependency>
             <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-executor-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-filestore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-keystore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-mapstore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-message-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-model-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-app-grpc</artifactId>
+            <artifactId>spals-appbuilder-app-examples</artifactId>
         </dependency>
         <dependency>
             <groupId>org.hamcrest</groupId>
@@ -90,46 +46,6 @@
 
     <build>
         <plugins>
-            <plugin>
-                <groupId>org.xolstice.maven.plugins</groupId>
-                <artifactId>protobuf-maven-plugin</artifactId>
-                <configuration>
-                    <protocArtifact>com.google.protobuf:protoc:${protobuf.version}:exe:${os.detected.classifier}</protocArtifact>
-                    <outputDirectory>target/generated-sources</outputDirectory>
-                    <clearOutputDirectory>false</clearOutputDirectory>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>generate-protobuf-java</id>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                    </execution>
-                    <execution>
-                        <id>generate-protobuf-grpc-java</id>
-                        <goals>
-                            <goal>compile-custom</goal>
-                        </goals>
-                        <configuration>
-                            <pluginId>grpc-java</pluginId>
-                            <pluginArtifact>
-                                io.grpc:protoc-gen-grpc-java:${grpc.version}:exe:${os.detected.classifier}
-                            </pluginArtifact>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>generate-protobuf-grpc-jersey</id>
-                        <goals>
-                            <goal>compile-custom</goal>
-                        </goals>
-                        <configuration>
-                            <pluginId>grpc-jersey</pluginId>
-                            <pluginArtifact>com.xorlev.grpc-jersey:protoc-gen-jersey:${grpc.jersey.version}:exe:${os.detected.classifier}</pluginArtifact>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
